+++ tests/test_colorize.py
@@ -1096,6 +1096,9 @@ def test_keyword_collection(invoke, assert_output_regex):
         r"\x1b\[94m\x1b\[1m\x1b\[4mExtra commands:\x1b\[0m\n"
         r"  \x1b\[36mcommand3\x1b\[0m\n"
         r"  \x1b\[36mcommand4\x1b\[0m  \x1b\[93m\x1b\[1m\(DEPRECATED\)\x1b\[0m\n"
+        r"\n"
+        r"\x1b\[94m\x1b\[1m\x1b\[4mOther commands:\x1b\[0m\n"
+        r"  \x1b\[36mhelp\x1b\[0m +Show help for a command\.\n"
     )
 
     result = invoke(color_cli1, "--help", color=True)